% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summary statistics of numeric variable}
\usage{
summary(data, x, y)
}
\arguments{
\item{data}{The dataset which need to be explore.
This parameter is named \code{data} because it tells the user intuitively that
we need a dataset as the first parameter.}

\item{x}{The numeric column name, which need to be summarized}

\item{y}{Another column name, the dataset can be grouped by this variable}
}
\value{
summarise_result: A list which contains the mean, range, median of numeric variable
}
\description{
This function do the summary statistics of one given numerical variable
across the groups of another given variable in a given dataset.
}
\examples{
result1 <- summary(apt_building, year_built, storey_height)
result2 <- summary(penguins, body_mass_g, species)
result3 <- summary(gapminder, lifeExp, continent)

}
